3주차 필수 과제

+++ src/main/java/org/sopt/service/PostService.java
@@ -2,11 +2,14 @@ package org.sopt.service;
 
 import jakarta.transaction.Transactional;
 import org.sopt.domain.Post;
+import org.sopt.domain.User;
 import org.sopt.dto.request.PostCreateRequest;
 import org.sopt.dto.request.PostPatchRequest;
+import org.sopt.dto.response.PostDetailResponse;
 import org.sopt.dto.response.PostListResponse;
 import org.sopt.dto.response.PostResponse;
 import org.sopt.repository.PostRepository;
+import org.sopt.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.net.URI;
@@ -16,32 +19,30 @@ import java.util.List;
 @Service
 public class PostService {
     private final PostRepository postRepository;
+    private final UserService userService;
 
-    public PostService(PostRepository postRepository) {
+    public PostService(PostRepository postRepository, UserService userService) {
         this.postRepository = postRepository;
+        this.userService = userService;
     }
 
     @Transactional
-    public URI createdPost(PostCreateRequest request) {
-        LocalDateTime recentCreatedTime = null;
+    public URI createdPost(PostCreateRequest request, Long userId) {
+        User currentUser = userService.getUserById(userId);
 
         if(postRepository.existsByTitle(request.title()))
             throw new IllegalArgumentException("중복된 제목을 가진 게시물이 존재합니다.");
 
-        if(!postTimeCheck(recentCreatedTime))
-            throw new IllegalStateException("최근 게시물을 작성한지 3분이 지나지 않았습니다.");
-
-        Post post = new Post(request.title());
+        Post post = new Post(request.title(), request.title(), currentUser);
 
         postRepository.save(post);
-        recentCreatedTime = LocalDateTime.now();
 
         return URI.create("/post/" + post.getId());
     }
 
     @Transactional
     public PostListResponse getAllPosts() {
-        List<PostResponse> postList = postRepository.findAll()
+        List<PostResponse> postList = postRepository.findAllByOrderByCreatedAtDesc()
                 .stream()
                 .map(PostResponse::from)
                 .toList();
@@ -50,25 +51,33 @@ public class PostService {
     }
 
     @Transactional
-    public PostResponse getPostById(Long postId) {
+    public PostDetailResponse getPostById(Long postId) {
         Post post = getPost(postId);
 
-        return PostResponse.from(post);
+        return PostDetailResponse.from(post);
     }
 
     @Transactional
-    public void deletePostById(Long id) {
+    public void deletePostById(Long id, Long userId) {
         Post post = getPost(id);
+        User currentUser = userService.getUserById(userId);
 
+        if(currentUser == post.getUser()) {
+            throw new IllegalArgumentException("삭제할 권한이 없는 사용자입니다.");
+        }
         postRepository.delete(post);
     }
-
     @Transactional
-    public PostResponse updatePostTitle(Long postId, PostPatchRequest request) {
+    public PostDetailResponse updatePost(Long userId, Long postId, PostPatchRequest request) {
+        User currentUser = userService.getUserById(userId);
         Post post = getPost(postId);
 
-        post.updateTitle(request.newTitle());
-        return PostResponse.from(post);
+        if(currentUser == post.getUser()) {
+            throw new IllegalArgumentException("수정할 권한이 없는 사용자입니다.");
+        }
+
+        post.updateTitleAndContent(request.title(), request.content());
+        return PostDetailResponse.from(post);
     }
 
     @Transactional
@@ -78,13 +87,6 @@ public class PostService {
                 .toList();
 
         return PostListResponse.of(postList);
-    }
-
-    private Boolean postTimeCheck(LocalDateTime recentCreatedTime) {
-        LocalDateTime threeMinutesAgo = LocalDateTime.now().minusMinutes(3);
-        if(recentCreatedTime != null && threeMinutesAgo.isBefore(recentCreatedTime))
-            return false;
-        return true;
     }
 
     private Post getPost(Long postId) {
